Path C: Composition in base, Micro::Attributes.new factory, hash API
Following the brainstorm, `Micro::Entity` is removed entirely.
Composition behavior (block-form attribute, hash → child-instance
coercion, deep validation bubbling, `__validate_nested_entities__`)
now lives inside `Micro::Attributes` itself — every includer gets it.

Three surface-level additions to the API, all on `Micro::Attributes`:

1. **Composition baked in.** `Micro::Attributes.included(base)` now
   `prepend`s `Composition::Coercion` and `include`s
   `Composition::Instance` on the base. `Macros#attribute` /
   `#attribute!` accept a `&block` and build the inline class via
   `__micro_attributes_build_inline_class__`, reusing the host's
   stored `with(...)` module so the inline child inherits the same
   feature mix. The `with(...)` class macro is also moved to Macros.

2. **`Micro::Attributes.new(options = {}, &block)`** — `Struct.new`-
   style factory. Returns a fresh class wired with `with(...)` on top
   of the preset `{ initialize: true, accept: true }`. Hash API only;
   each option is overridable per-key.

3. **Hash-style `Micro::Attributes.with(...)`** — `split_strict_hash`
   in features.rb is generalized into a per-entry normalizer that
   handles `initialize: true|:strict`, `accept: true|:strict`,
   `diff: true`, `keys_as: :symbol|:string|:indifferent`, and
   `active_model: :validations`. The legacy positional symbol API
   stays fully supported; both forms can be mixed.

Other infrastructure changes:

- `Features::ActiveModelValidations.included` now triggers the
  `validate :__validate_nested_entities__` registration for any
  `Micro::Attributes` includer (was: only `Micro::Entity` subclasses).
- `Features.with(...)` builders memoize a reference to the produced
  `With::*` module on every including class via the new
  `@__micro_attributes_with_module__` ivar. `Micro::Attributes`'
  `inherited` hook copies that ivar to subclasses so deep inheritance
  trees still build inline children with the right feature mix.
- `lib/micro/entity.rb` is gone. The compat-class shim was deemed
  unnecessary because `Micro::Entity` was never released.

Tests:

- `test/micro/entity_test.rb` → `test/micro/attributes/composition_test.rb`,
  rewritten to use `Micro::Attributes.new` / `Micro::Attributes.with(...)`
  instead of `< Micro::Entity`. Every behavior from the old Entity
  tests is preserved (block-form, multi-level isolation, ordering,
  E4 stable inline-class naming, strict, deep nesting, AM bubbling,
  mixed-tree AM, with-macro layering).
- `test/micro/entity_matrix_test.rb` → `test/micro/attributes/composition_matrix_test.rb`,
  rewritten so the matrix axes use the hash API directly. One bonus:
  the inline block-form child now propagates the outer's feature mix
  (the old E1 "doesn't propagate" tradeoff resolves naturally because
  the inline class includes the same With::* module).

README:

- Removed the entire `# Micro::Entity` section.
- Added `# Composition` section with subsections for
  `Micro::Attributes.new`, the hash API for `.with`, nested via
  `accept:`, block-form, and deep nesting + validation bubbling.
- ToC updated.

CHANGELOG:

- Rewrote the 3.1.0 `Added` section to lead with the
  composition-in-base story, the `.new` factory, and the hash API.
- Kept both `Changed` callouts (multi-key strict hash shift; Accept
  visibility-gate consequences for round-trip + diff).
- Kept all `Fixed` bullets from earlier in this PR.

Baseline: 221 runs / 3215 assertions / 0 failures.
rails-8-1: 230 runs / 4784 assertions / 0 failures.
