+++ test/micro/attributes/composition_matrix_test.rb
@@ -1,7 +1,6 @@
 require 'test_helper'
-require 'micro/entity'
 
-ENTITY_MATRIX_HAS_ACTIVEMODEL =
+COMPOSITION_MATRIX_HAS_ACTIVEMODEL =
   begin
     require 'active_model'
     true
@@ -9,81 +8,78 @@ ENTITY_MATRIX_HAS_ACTIVEMODEL =
     false
   end
 
-# Matrix test for Micro::Entity across every supported feature combination.
+# Matrix test exercising composition behavior — block-form attribute,
+# hash coercion, deep validation — across every supported feature
+# combination of `Micro::Attributes.new(...)`.
 #
 # Axes:
-# - Base:          Micro::Entity  vs  Micro::Entity::Strict
-# - Key access:    default (string/indifferent)  vs  KeysAsSymbol
-# - Validations:   none  vs  ActiveModelValidations  (gated on activemodel being loadable)
+# - Strict:        loose vs strict (initialize: :strict + accept: :strict)
+# - Key access:    default (string/indifferent) vs KeysAsSymbol
+# - Validations:   none vs ActiveModelValidations (gated on activemodel being loadable)
 #
 # Invariants verified for every cell:
 # - Basic instantiation from a hash works.
-# - `accept:` on a nested Entity subclass coerces a hash to the entity.
-# - An entity instance passes through (`accept` check succeeds, same object).
-# - The block form of `attribute` defines an anonymous nested entity at runtime.
+# - `accept:` on a nested class coerces a hash to the entity.
+# - An instance passes through (`accept` check succeeds, same object).
+# - The block form of `attribute` defines an anonymous nested at runtime.
 # - `#with_attribute` produces a new instance (immutability).
-class Micro::EntityMatrixTest < Minitest::Test
-  BASE_AXIS = [
-    [:loose,  Micro::Entity,         false],
-    [:strict, Micro::Entity::Strict, true],
+class Micro::Attributes::CompositionMatrixTest < Minitest::Test
+  STRICT_AXIS = [
+    [:loose,  false, {}],
+    [:strict, true,  { initialize: :strict, accept: :strict }],
   ].freeze
 
   KEYS_AXIS = [
-    [:string, nil,                                            ->(name) { name.to_s }],
-    [:symbol, Micro::Attributes.with(:keys_as_symbol),        ->(name) { name.to_sym }],
+    [:string, {},                  ->(name) { name.to_s }],
+    [:symbol, { keys_as: :symbol }, ->(name) { name.to_sym }],
   ].freeze
 
   AM_AXIS =
-    if ENTITY_MATRIX_HAS_ACTIVEMODEL
+    if COMPOSITION_MATRIX_HAS_ACTIVEMODEL
       [
-        [:no_am, nil],
-        [:am,    Micro::Attributes.with(:activemodel_validations)],
+        [:no_am, {}],
+        [:am,    { active_model: :validations }],
       ].freeze
     else
-      [[:no_am, nil]].freeze
+      [[:no_am, {}]].freeze
     end
 
   # Build a `[outer_class, nested_class]` pair for a given matrix cell.
-  # Each axis contributes one feature module that we mix into the entity.
   # Classes are assigned to constants because ActiveModel uses `Class#name`
   # when rendering validation messages (anonymous classes raise).
-  def self.build_cell(label, base, keys_module, am_module)
+  def self.build_cell(label, strict_opts, keys_opts, am_opts)
     const_label = label.tr('/', '_').gsub(/\W+/, '_').sub(/^_+/, '').sub(/_+$/, '')
 
-    nested = Class.new(base) do
-      include keys_module if keys_module
-      include am_module if am_module
+    base_opts = strict_opts.merge(keys_opts).merge(am_opts)
+    has_am = !am_opts.empty?
 
+    nested = Micro::Attributes.new(**base_opts) do
       attribute :value, accept: Integer
     end
     const_set("Nested_#{const_label}", nested)
 
-    outer = Class.new(base) do
-      include keys_module if keys_module
-      include am_module if am_module
-
+    outer = Micro::Attributes.new(**base_opts) do
       attribute :name,   accept: String
       attribute :nested, accept: nested
 
-      # Block form must always work — anon nested entity defined at runtime.
+      # Block form must always work — anon nested defined at runtime.
       attribute :inline do
         attribute :label, accept: String
       end
-
-      # AM cells get a real validation rule so we can exercise `valid?`,
-      # `errors`, and the AM→attributes_errors bridge in the assertions.
-      validates :name, presence: true if am_module
     end
+    # AM cells get a real validation rule so we can exercise `valid?`,
+    # `errors`, and the AM→attributes_errors bridge in the assertions.
+    outer.validates(:name, presence: true) if has_am
     const_set("Outer_#{const_label}", outer)
 
     [outer, nested]
   end
 
-  CELLS = BASE_AXIS.flat_map do |base_label, base, strict|
-    KEYS_AXIS.flat_map do |keys_label, keys_module, key_for|
-      AM_AXIS.map do |am_label, am_module|
-        label = "#{base_label}/#{keys_label}/#{am_label}"
-        outer, nested = build_cell(label, base, keys_module, am_module)
+  CELLS = STRICT_AXIS.flat_map do |strict_label, strict, strict_opts|
+    KEYS_AXIS.flat_map do |keys_label, keys_opts, key_for|
+      AM_AXIS.map do |am_label, am_opts|
+        label = "#{strict_label}/#{keys_label}/#{am_label}"
+        outer, nested = build_cell(label, strict_opts, keys_opts, am_opts)
         {
           label: label,
           strict: strict,
@@ -96,7 +92,7 @@ class Micro::EntityMatrixTest < Minitest::Test
   end.freeze
 
   def test_matrix_has_expected_size
-    expected = ENTITY_MATRIX_HAS_ACTIVEMODEL ? 8 : 4
+    expected = COMPOSITION_MATRIX_HAS_ACTIVEMODEL ? 8 : 4
     assert_equal(expected, CELLS.size, 'unexpected matrix cardinality')
   end
 
@@ -111,18 +107,17 @@ class Micro::EntityMatrixTest < Minitest::Test
       obj = outer.new(attrs)
 
       assert_equal('Rodrigo', obj.name, "name reader (#{cell[:label]})")
-      assert_kind_of(cell[:nested], obj.nested, "nested coerced to entity (#{cell[:label]})")
+      assert_kind_of(cell[:nested], obj.nested, "nested coerced (#{cell[:label]})")
       assert_equal(42, obj.nested.value, "nested value (#{cell[:label]})")
-      assert_kind_of(::Micro::Entity, obj.inline, "inline coerced to entity (#{cell[:label]})")
+      assert(obj.inline.class.include?(::Micro::Attributes), "inline includes Attributes (#{cell[:label]})")
       assert_equal('hi', obj.inline.label, "inline label (#{cell[:label]})")
 
-      # attributes hash keys follow the KeysAsSymbol axis.
       assert(obj.attributes.key?(key.call(:name)), "name in attributes (#{cell[:label]})")
       refute_predicate(obj, :attributes_errors?, "no accept errors (#{cell[:label]})")
     end
   end
 
-  def test_entity_instance_passes_through_without_re_coercion
+  def test_instance_passes_through_without_re_coercion
     CELLS.each do |cell|
       pre_built = cell[:nested].new(value: 7)
 
@@ -132,7 +127,7 @@ class Micro::EntityMatrixTest < Minitest::Test
         inline: { label: 'L' },
       )
 
-      assert_same(pre_built, obj.nested, "entity instance not re-coerced (#{cell[:label]})")
+      assert_same(pre_built, obj.nested, "instance not re-coerced (#{cell[:label]})")
     end
   end
 
@@ -153,11 +148,10 @@ class Micro::EntityMatrixTest < Minitest::Test
           nested: { value: 1 },
           inline: { label: :not_a_string },
         )
-        assert_predicate(obj.inline, :attributes_errors?, "inline has the detail (#{cell[:label]})")
+        assert_predicate(obj.inline, :attributes_errors?, "inline has detail (#{cell[:label]})")
 
-        # Deep-nesting bubble: the outer mirrors descendant invalidity via
-        # a `'is invalid'` marker (the leaf retains the full message).
-        assert_predicate(obj, :attributes_errors?, "outer mirrors inline invalidity (#{cell[:label]})")
+        # Deep bubble: outer mirrors descendant invalidity via marker.
+        assert_predicate(obj, :attributes_errors?, "outer mirrors inline (#{cell[:label]})")
         key_for_inline = cell[:key_for].call(:inline)
         assert_equal('is invalid', obj.attributes_errors[key_for_inline],
                      "outer carries the marker (#{cell[:label]})")
@@ -208,36 +202,23 @@ class Micro::EntityMatrixTest < Minitest::Test
 
       updated = obj.with_attribute(:name, 'B')
 
-      refute_same(obj, updated, "with_attribute returns new instance (#{cell[:label]})")
-      assert_equal('A', obj.name,     "original unchanged (#{cell[:label]})")
-      assert_equal('B', updated.name, "updated reflects change (#{cell[:label]})")
+      refute_same(obj, updated, "with_attribute returns new (#{cell[:label]})")
+      assert_equal('A', obj.name, cell[:label])
+      assert_equal('B', updated.name, cell[:label])
     end
   end
 
-  def test_diff_is_available_across_matrix
-    CELLS.reject { |c| c[:strict] }.each do |cell|
-      a = cell[:outer].new(name: 'A', nested: { value: 1 }, inline: { label: 'L' })
-      b = a.with_attribute(:name, 'B')
+  # -- ActiveModel-specific assertions (only when AM is loadable) -----------
 
-      diff = a.diff_attributes(b)
-      key = cell[:key_for]
-
-      assert(diff.changed?(key.call(:name)), "diff sees name change (#{cell[:label]})")
-      refute(diff.changed?(key.call(:nested)), "nested unchanged (#{cell[:label]})")
-    end
-  end
-
-  # -- ActiveModel-specific assertions (only when AM is loadable) ---------------------------
-
-  if ENTITY_MATRIX_HAS_ACTIVEMODEL
+  if COMPOSITION_MATRIX_HAS_ACTIVEMODEL
     AM_CELLS = CELLS.select { |c| c[:outer].include?(Micro::Attributes::Features::ActiveModelValidations) }
 
-    def test_activemodel_valid_predicate_works_with_entity
+    def test_am_valid_predicate_works
       AM_CELLS.each do |cell|
         valid = cell[:outer].new(name: 'Rodrigo', nested: { value: 1 }, inline: { label: 'L' })
         assert_predicate(valid, :valid?, "valid (#{cell[:label]})")
 
-        next if cell[:strict] # strict raises on accept errors before validations are interesting
+        next if cell[:strict]
 
         invalid = cell[:outer].new(name: '', nested: { value: 1 }, inline: { label: 'L' })
         refute_predicate(invalid, :valid?, "invalid blank name (#{cell[:label]})")
@@ -245,9 +226,7 @@ class Micro::EntityMatrixTest < Minitest::Test
       end
     end
 
-    def test_activemodel_errors_merge_into_attributes_errors
-      # When accept and activemodel_validations are combined, AM errors are surfaced via
-      # attributes_errors (loose cells). Strict cells raise instead.
+    def test_am_errors_merge_into_attributes_errors
       AM_CELLS.reject { |c| c[:strict] }.each do |cell|
         obj = cell[:outer].new(name: '', nested: { value: 1 }, inline: { label: 'L' })
